# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "User micro service",
    "version": "1.0.0",
    "title": "User Service"
  },
  "host": "127.0.0.1:8081",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all the users",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/usersResponse"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Logs a user into the service",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "description": "Login Request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/loginRequest"
            },
            "name": "User"
          }
        ],
        "responses": {
          "200": {
            "description": "Logged in'",
            "schema": {
              "$ref": "#/definitions/loginResponse"
            }
          },
          "401": {
            "description": "Not logged in"
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Logout a user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Logged in'"
          },
          "401": {
            "description": "Not logged in"
          }
        }
      }
    },
    "/user/{username}/exists": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Checks if the username exists",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Users username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "True or False",
            "schema": {
              "$ref": "#/definitions/usernameResponse"
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Returns the current user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Found",
            "schema": {
              "$ref": "#/definitions/userResponse"
            }
          },
          "404": {
            "description": "You are not logged in"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Creates a user",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "description": "Create Request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userPostRequest"
            },
            "name": "User"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Updates a user",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "description": "Update Request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userPutRequest"
            },
            "name": "User"
          }
        ],
        "responses": {
          "200": {
            "description": "User updated succesfully!"
          },
          "400": {
            "description": "Error"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Found",
            "schema": {
              "$ref": "#/definitions/userResponse"
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "definitions": {
    "loginRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "username"
        },
        "password": {
          "type": "string",
          "example": "password"
        }
      },
      "required": [
        "username",
        "password"
      ]
    },
    "loginResponse": {
      "type": "object",
      "properties": {
        "api_key": {
          "type": "string",
          "example": "api_key"
        },
        "message": {
          "type": "string",
          "example": "Logged In"
        }
      }
    },
    "usernameResponse": {
      "type": "object",
      "properties": {
        "exists": {
          "type": "boolean",
          "example": "True"
        }
      }
    },
    "userPostRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "username"
        },
        "password": {
          "type": "string",
          "example": "password"
        },
        "first_name": {
          "type": "string",
          "example": "firstname"
        },
        "last_name": {
          "type": "string",
          "example": "lastname"
        },
        "email": {
          "type": "string",
          "example": "abc@gmail.com"
        }
      },
      "required": [
        "username",
        "password",
        "first_name",
        "email",
        "last_name"
      ]
    },
    "userPutRequest": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "firstname"
        },
        "last_name": {
          "type": "string",
          "example": "lastname"
        },
        "email": {
          "type": "string",
          "example": "abc@gmail.com"
        }
      },
      "required": [
        "first_name",
        "last_name",
        "email"
      ]
    },
    "userResponse": {
      "type": "object",
      "properties": {
        "date_added": {
          "type": "string",
          "example": "2020-12-04T14:39:25"
        },
        "date_updated": {
          "type": "string",
          "example": "2020-12-04T14:39:25"
        },
        "first_name": {
          "type": "string",
          "example": "firstname"
        },
        "email": {
          "type": "string",
          "example": "abc@gmail.com"
        },
        "id": {
          "type": "integer",
          "example": "1"
        },
        "last_name": {
          "type": "string",
          "example": "lastname"
        },
        "username": {
          "type": "string",
          "example": "username"
        }
      }
    },
    "usersResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "$ref": "#/definitions/userResponse"
      }
    }
  }
}